+++ include/ylt/metric/gauge.hpp
